[Evol] implementation of table processing
 Table processing requires some consequent refactoring as they appear at
the same level of paragraph and imply some recursive treatments as table
cells are considered Body just like headers, footers and documents
themselves. 
  
  - implementation of the table parsing 
  - implementation of the static table preservation
  
  The result of this commit can be used to produce iterated tables
because words make consecutive tables appear to be the same. However,
the gd:tablemerge tag must be implemented for a full support of tables
and being able to insert legends.