Issue #82: Make the core runtime independent from the Workspace
And even independant from java.io.File which is even more interesting!

By:
- replace all refs to workspace/java.io.File types by the emf URI
- removed all the references to "project location" from the runtime
- move to Java 7 to use the try{} with resource and fix some resource
leaks along the way
- changing the way the template file name and result file name
attributes in the genconf model are interpreted, instead of being path
starting from the current project they are now path starting from the
current model. /!\ This is an impacting change which deserves a
proeminent message in the release notes.
- update the genconf models by adapting the template and result name
paths

https://github.com/ObeoNetwork/M2Doc/issues/82