Issue #82 Apply EMF pattern to get o.o.m2doc plugin usable as simple Jar
This pattern is applied all over in EMF and helps in bringing standalone
support to pre-existing plugins without splitting it in two. Most
notably:

- change the implementation of the org.obeonetwork.m2doc activator to
extract the Eclipse specific part into a specific "Implementation"
class.
- adapt the Manifest.MF to declare this class as the OSGi activator
implementation and change all the calls which were relying to OSGi
having initialized something to direct access to an INSTANCE singleton.

The main implication from an API user point of view is that
M2DocPlugin.INSTANCE is always there and logging through it will either
log through Sys.out or through the Eclipse logger depending on the
current environment.

Same treatment is applied on the m2doc.tests plugins and the
DeclaredProviderListenerTest test is now clarifying its assumption that
the eclipse platform is running. Tests can now be run as JUnit tests
(and not only JUnit Plugin Tests).