[Fix] fix #118
M2Doc cannot be launched on a genconf in CDO #118

The solution this fix provides consists in having a distinct command
handler in the sirius plug-in that is used on objects in a CDO
repositories (the URI is used to decide which handler to use). The
session is used to find out a platform resource in the session to get a
project from.