Add an attribute in genconf to know which representation file to load
This is a first step to remove the limitation of having to put the
genconf as part of the modeling project it is using. 

By removing the assumption that the genconf model is "part of a Sirius
session" this also allow to launch the document generation right from
the genconf file and not from the "Generation" instance which got loaded
by Sirius in some way.