Adapt the DiagramProvider API to pass the required resourceset
This specify to the diagram provider where to search for
representations. In the case of the Sirius one it allows to retrive the
session without relying on the assumption -- no longer true -- that the
Generation instance is part of the Session.

This commit also adapts the Sirius integration tests which no longer
needs the specific initialization as it is now taken care of by the
m2doc runtime through the Sirius integration plugin.