Fix resolution of the template URI in the case where a genconf is used.
The template URI was computed as a *relative* URI, based on the path
entered by the user in the genconf.
During the resolution of inserted images, this relative URI was used as
a the base URI. This failed because the base URI for a resolution should
be hierarchical or absolute.

The template URI is now computed, when possible, a resolution of the
path entered by the user in the genconf, against the URI of the genconf
itself.

Change-Id: I54dd37b5fba31c746a34c5e9180f2f4f2e1b5885
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>